refactor(integrations): split blocker and WakaTime internals
Extract hosts-file command helpers, block rules, and blocker tests into dedicated blocker modules.

Split WakaTime config, queue, transport, and tests so tracking concerns are isolated behind the existing facade.
