+++ src/blocker.rs
@@ -1,30 +1,30 @@
 #[cfg(test)]
 use std::cell::RefCell;
 use std::collections::HashSet;
-use std::fs::{self, OpenOptions};
+use std::fs;
 use std::io;
-use std::path::{Path, PathBuf};
-use std::process::Command;
+use std::path::Path;
 
-#[cfg(target_os = "windows")]
-const HOSTS_FILE: &str = r"C:\Windows\System32\drivers\etc\hosts";
-#[cfg(not(target_os = "windows"))]
-const HOSTS_FILE: &str = "/etc/hosts";
 const BLOCK_MARKER_START: &str = "# focustime-block-start";
 const BLOCK_MARKER_END: &str = "# focustime-block-end";
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum HostsWriteFailStep {
-    Snapshot,
-    StageWrite,
-    AfterReplace,
-    RollbackRestore,
-}
+mod command;
+mod hosts;
+mod rules;
+
+pub use command::CommandBlockingBackend;
+use command::{apply_command_backend, command_backend_diagnostics, preview_from_command};
+use hosts::{HOSTS_FILE, atomic_write_hosts_to_path, flush_dns_cache, hosts_file_diagnostics_for};
+#[cfg(test)]
+use hosts::{HostsWriteFailStep, set_test_hosts_write_fail_steps};
+
+#[cfg(test)]
+pub use rules::normalize_domain_host;
+pub use rules::{SiteValidationError, domain_rule_matches_host, normalize_domain_rule};
 
 #[cfg(test)]
 thread_local! {
     static TEST_LAST_BLOCKING_ACTION: RefCell<Option<&'static str>> = const { RefCell::new(None) };
-    static TEST_HOSTS_WRITE_FAIL_STEPS: RefCell<Vec<HostsWriteFailStep>> = const { RefCell::new(Vec::new()) };
 }
 
 #[cfg(test)]
@@ -37,13 +37,6 @@ fn record_test_blocking_action(action: &'static str) {
 #[cfg(test)]
 pub(crate) fn take_test_blocking_action() -> Option<&'static str> {
     TEST_LAST_BLOCKING_ACTION.with(|slot| slot.borrow_mut().take())
-}
-
-#[cfg(test)]
-fn set_test_hosts_write_fail_steps(steps: &[HostsWriteFailStep]) {
-    TEST_HOSTS_WRITE_FAIL_STEPS.with(|slot| {
-        *slot.borrow_mut() = steps.to_vec();
-    });
 }
 
 pub struct SiteBlocker {
@@ -72,147 +65,6 @@ impl HostsFileDiagnostics {
     pub fn can_write(&self) -> bool {
         self.write_error.is_none()
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SiteValidationError {
-    EmptyHostname,
-    MissingHostname,
-    ContainsWhitespace,
-    InvalidCharacter,
-    InvalidLabel,
-    MultipleHostnames,
-}
-
-impl SiteValidationError {
-    pub fn message(self) -> &'static str {
-        match self {
-            SiteValidationError::EmptyHostname => "empty hostname",
-            SiteValidationError::MissingHostname => "missing hostname",
-            SiteValidationError::ContainsWhitespace => "contains whitespace",
-            SiteValidationError::InvalidCharacter => "contains invalid characters",
-            SiteValidationError::InvalidLabel => "invalid hostname format",
-            SiteValidationError::MultipleHostnames => "multiple hostnames not allowed",
-        }
-    }
-}
-
-pub fn normalize_domain_rule(input: &str) -> Result<String, SiteValidationError> {
-    normalize_domain_like_input(input, true)
-}
-
-pub fn normalize_domain_host(input: &str) -> Result<String, SiteValidationError> {
-    normalize_domain_like_input(input, false)
-}
-
-pub fn domain_rule_matches_host(rule: &str, host: &str) -> bool {
-    let Ok(rule) = normalize_domain_rule(rule) else {
-        return false;
-    };
-    let Ok(host) = normalize_domain_host(host) else {
-        return false;
-    };
-    if let Some(suffix) = rule.strip_prefix("*.") {
-        return host.ends_with(&format!(".{suffix}"));
-    }
-    host == rule
-}
-
-fn normalize_domain_like_input(
-    input: &str,
-    allow_wildcard_prefix: bool,
-) -> Result<String, SiteValidationError> {
-    let hostname = extract_hostname_candidate(input)?;
-    let (hostname, wildcard_prefix) = extract_wildcard_prefix(hostname, allow_wildcard_prefix)?;
-    let hostname = strip_trailing_root_dots(strip_numeric_port(hostname)?);
-    if hostname.is_empty() {
-        return Err(SiteValidationError::MissingHostname);
-    }
-    validate_domain_host(&hostname)?;
-    if wildcard_prefix && !hostname.contains('.') {
-        return Err(SiteValidationError::InvalidLabel);
-    }
-    if wildcard_prefix {
-        Ok(format!("*.{hostname}"))
-    } else {
-        Ok(hostname)
-    }
-}
-
-fn extract_hostname_candidate(input: &str) -> Result<String, SiteValidationError> {
-    let mut hostname = input.trim().to_lowercase();
-    if hostname.is_empty() {
-        return Err(SiteValidationError::EmptyHostname);
-    }
-    if let Some(sep) = hostname.find("://") {
-        hostname = hostname[sep + 3..].to_string();
-    }
-    if let Some(pos) = hostname.find(['/', '?', '#']) {
-        hostname.truncate(pos);
-    }
-    if let Some(at_pos) = hostname.rfind('@') {
-        hostname = hostname[at_pos + 1..].to_string();
-    }
-    Ok(hostname)
-}
-
-fn extract_wildcard_prefix(
-    hostname: String,
-    allow_wildcard_prefix: bool,
-) -> Result<(String, bool), SiteValidationError> {
-    if let Some(stripped) = hostname.strip_prefix("*.") {
-        if !allow_wildcard_prefix {
-            return Err(SiteValidationError::InvalidCharacter);
-        }
-        return Ok((stripped.to_string(), true));
-    }
-    if let Some(stripped) = hostname.strip_prefix('.') {
-        return Ok((stripped.to_string(), allow_wildcard_prefix));
-    }
-    Ok((hostname, false))
-}
-
-fn strip_numeric_port(mut hostname: String) -> Result<String, SiteValidationError> {
-    if let Some(colon_pos) = hostname.rfind(':') {
-        let port = &hostname[colon_pos + 1..];
-        if hostname[..colon_pos].contains(':') || !port.chars().all(|c| c.is_ascii_digit()) {
-            return Err(SiteValidationError::InvalidLabel);
-        }
-        hostname.truncate(colon_pos);
-    }
-    Ok(hostname)
-}
-
-fn strip_trailing_root_dots(mut hostname: String) -> String {
-    while hostname.ends_with('.') {
-        hostname.pop();
-    }
-    hostname
-}
-
-fn validate_domain_host(hostname: &str) -> Result<(), SiteValidationError> {
-    if hostname.chars().any(char::is_whitespace) {
-        return Err(SiteValidationError::ContainsWhitespace);
-    }
-    if !hostname
-        .chars()
-        .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '.' || c == '-')
-    {
-        return Err(SiteValidationError::InvalidCharacter);
-    }
-    if hostname.starts_with('.')
-        || hostname.ends_with('.')
-        || hostname.contains("..")
-        || hostname.len() > 253
-    {
-        return Err(SiteValidationError::InvalidLabel);
-    }
-    for label in hostname.split('.') {
-        if label.is_empty() || label.len() > 63 || label.starts_with('-') || label.ends_with('-') {
-            return Err(SiteValidationError::InvalidLabel);
-        }
-    }
-    Ok(())
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -298,27 +150,6 @@ impl BlockingBackendPolicy {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct CommandBlockingBackend {
-    pub block_command: String,
-    pub unblock_command: String,
-    pub diagnostics_command: String,
-}
-
-impl CommandBlockingBackend {
-    pub fn normalized(&self) -> Self {
-        Self {
-            block_command: self.block_command.trim().to_string(),
-            unblock_command: self.unblock_command.trim().to_string(),
-            diagnostics_command: self.diagnostics_command.trim().to_string(),
-        }
-    }
-
-    pub fn is_configured(&self) -> bool {
-        !self.block_command.trim().is_empty() && !self.unblock_command.trim().is_empty()
-    }
-}
-
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct BlockingBackendStatus {
     pub policy: BlockingBackendPolicy,
@@ -393,16 +224,7 @@ impl SiteBlocker {
     }
 
     pub fn command_backend_diagnostics(&self) -> io::Result<()> {
-        if self.command_backend.diagnostics_command.trim().is_empty() {
-            if self.command_backend.is_configured() {
-                return Ok(());
-            }
-            return Err(io::Error::new(
-                io::ErrorKind::NotFound,
-                "command backend block/unblock commands are not fully configured",
-            ));
-        }
-        run_shell_command(&self.command_backend.diagnostics_command)
+        command_backend_diagnostics(&self.command_backend)
     }
 
     pub fn add_site(&mut self, site: String) {
@@ -660,70 +482,11 @@ impl SiteBlocker {
     }
 
     fn apply_command_backend(&self, intent: BlockingIntent) -> io::Result<()> {
-        let template = match intent {
-            BlockingIntent::Block => self.command_backend.block_command.as_str(),
-            BlockingIntent::Unblock => self.command_backend.unblock_command.as_str(),
-        };
-        let template = template.trim();
-        if template.is_empty() {
-            return Err(io::Error::new(
-                io::ErrorKind::NotFound,
-                format!(
-                    "command backend `{}` command is not configured",
-                    blocking_intent_id(intent)
-                ),
-            ));
-        }
-        let rendered = render_command_template(template, &self.sites);
-        run_shell_command(&rendered)
+        apply_command_backend(&self.command_backend, intent, &self.sites)
     }
 
     fn preview_from_command(&self, intent: BlockingIntent) -> io::Result<BlockingPreview> {
-        let template = match intent {
-            BlockingIntent::Block => self.command_backend.block_command.as_str(),
-            BlockingIntent::Unblock => self.command_backend.unblock_command.as_str(),
-        }
-        .trim()
-        .to_string();
-        if template.is_empty() {
-            return Err(io::Error::new(
-                io::ErrorKind::NotFound,
-                format!(
-                    "command backend `{}` command is not configured",
-                    blocking_intent_id(intent)
-                ),
-            ));
-        }
-
-        let rendered = render_command_template(&template, &self.sites);
-        let action = match intent {
-            BlockingIntent::Block => {
-                if self.sites.is_empty() {
-                    BlockingPreviewAction::NoChange
-                } else {
-                    BlockingPreviewAction::Block
-                }
-            }
-            BlockingIntent::Unblock => BlockingPreviewAction::Unblock,
-        };
-        let effective_blocked_sites = if action == BlockingPreviewAction::Block {
-            self.sites.clone()
-        } else {
-            Vec::new()
-        };
-
-        Ok(BlockingPreview {
-            backend: BlockingBackendKind::Command,
-            backend_target: rendered.clone(),
-            attempted_backends: vec![BlockingBackendKind::Command],
-            fallback_used: false,
-            hosts_file_path: rendered,
-            action,
-            effective_blocked_sites,
-            would_change: action != BlockingPreviewAction::NoChange,
-            current_section: None,
-            next_section: None,
-        })
+        preview_from_command(&self.command_backend, intent, &self.sites)
     }
 
     fn apply_hosts_block(&self) -> io::Result<()> {
@@ -958,43 +721,6 @@ fn blocking_intent_id(intent: BlockingIntent) -> &'static str {
     }
 }
 
-fn render_command_template(template: &str, sites: &[String]) -> String {
-    let sites_csv = sites.join(",");
-    let sites_lines = sites.join("\n");
-    template
-        .replace("{sites_csv}", &sites_csv)
-        .replace("{sites_lines}", &sites_lines)
-        .replace("{site_count}", &sites.len().to_string())
-}
-
-fn run_shell_command(command: &str) -> io::Result<()> {
-    let status = shell_command(command).status()?;
-    if status.success() {
-        return Ok(());
-    }
-    Err(io::Error::other(format!(
-        "command exited with status {}",
-        status
-            .code()
-            .map(|code| code.to_string())
-            .unwrap_or_else(|| "unknown".to_string())
-    )))
-}
-
-#[cfg(target_os = "windows")]
-fn shell_command(command: &str) -> Command {
-    let mut cmd = Command::new("cmd");
-    cmd.args(["/C", command]);
-    cmd
-}
-
-#[cfg(not(target_os = "windows"))]
-fn shell_command(command: &str) -> Command {
-    let mut cmd = Command::new("sh");
-    cmd.args(["-c", command]);
-    cmd
-}
-
 fn line_ending_for(content: &str) -> &'static str {
     if content.contains("\r\n") {
         "\r\n"
@@ -1029,800 +755,5 @@ fn split_hostname_candidates(input: &str) -> Vec<String> {
         .collect()
 }
 
-fn hosts_file_diagnostics_for(path: &Path) -> HostsFileDiagnostics {
-    let read_error = fs::File::open(path).err().map(|error| error.to_string());
-    let write_error = probe_hosts_write_path(path)
-        .err()
-        .map(|error| error.to_string());
-    HostsFileDiagnostics {
-        path: path.display().to_string(),
-        read_error,
-        write_error,
-    }
-}
-
-#[cfg(target_os = "windows")]
-fn probe_hosts_write_path(path: &Path) -> io::Result<()> {
-    OpenOptions::new().append(true).open(path).map(|_| ())
-}
-
-#[cfg(not(target_os = "windows"))]
-fn probe_hosts_write_path(path: &Path) -> io::Result<()> {
-    let dir = path.parent().unwrap_or(Path::new("."));
-    let nanos = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|duration| duration.as_nanos())
-        .unwrap_or(0);
-    let probe = dir.join(format!(
-        ".focustime_hosts_probe_{}_{}",
-        std::process::id(),
-        nanos
-    ));
-    OpenOptions::new()
-        .write(true)
-        .create_new(true)
-        .open(&probe)?;
-    fs::remove_file(&probe)
-}
-
-enum HostsReplaceError {
-    NoMutation(io::Error),
-    #[cfg(target_os = "windows")]
-    NeedsRollback(io::Error),
-}
-
-/// Update the hosts file via staged replacement and rollback, deriving next
-/// content from the snapshotted source to avoid TOCTOU races.
-fn atomic_write_hosts_to_path<F>(hosts_path: &Path, build_content: F) -> io::Result<bool>
-where
-    F: FnOnce(&str) -> Option<String>,
-{
-    let staged_path = hosts_transaction_temp_path(hosts_path, "staged");
-    let snapshot_path = hosts_transaction_temp_path(hosts_path, "snapshot");
-
-    create_hosts_snapshot(hosts_path, &snapshot_path)?;
-    let snapshot_content = match fs::read_to_string(&snapshot_path) {
-        Ok(content) => content,
-        Err(error) => {
-            let _ = remove_file_if_exists(&staged_path);
-            let _ = remove_file_if_exists(&snapshot_path);
-            return Err(error);
-        }
-    };
-    let Some(content) = build_content(&snapshot_content) else {
-        let _ = remove_file_if_exists(&staged_path);
-        let _ = remove_file_if_exists(&snapshot_path);
-        return Ok(false);
-    };
-
-    if let Err(error) = write_staged_hosts_file(hosts_path, &staged_path, &content) {
-        let _ = remove_file_if_exists(&staged_path);
-        let _ = remove_file_if_exists(&snapshot_path);
-        return Err(error);
-    }
-
-    match replace_hosts_file(&staged_path, hosts_path) {
-        Ok(()) => {}
-        Err(HostsReplaceError::NoMutation(error)) => {
-            let _ = remove_file_if_exists(&staged_path);
-            let _ = remove_file_if_exists(&snapshot_path);
-            return Err(error);
-        }
-        #[cfg(target_os = "windows")]
-        Err(HostsReplaceError::NeedsRollback(error)) => {
-            return rollback_hosts_write(hosts_path, &snapshot_path, &staged_path, error)
-                .map(|()| true);
-        }
-    }
-
-    if let Err(error) = maybe_fail_hosts_write_step(HostsWriteFailStep::AfterReplace) {
-        return rollback_hosts_write(hosts_path, &snapshot_path, &staged_path, error)
-            .map(|()| true);
-    }
-
-    cleanup_temp_file_best_effort(&staged_path);
-    cleanup_temp_file_best_effort(&snapshot_path);
-    Ok(true)
-}
-
-fn rollback_hosts_write(
-    hosts_path: &Path,
-    snapshot_path: &Path,
-    staged_path: &Path,
-    update_error: io::Error,
-) -> io::Result<()> {
-    let rollback_result = restore_hosts_snapshot(hosts_path, snapshot_path);
-    let _ = remove_file_if_exists(staged_path);
-    let _ = remove_file_if_exists(snapshot_path);
-
-    match rollback_result {
-        Ok(()) => Err(io::Error::new(
-            update_error.kind(),
-            format!("failed to update hosts file: {update_error}; previous state restored"),
-        )),
-        Err(rollback_error) => Err(io::Error::other(format!(
-            "failed to update hosts file: {update_error}; rollback failed: {rollback_error}"
-        ))),
-    }
-}
-
-fn create_hosts_snapshot(hosts_path: &Path, snapshot_path: &Path) -> io::Result<()> {
-    maybe_fail_hosts_write_step(HostsWriteFailStep::Snapshot)?;
-    fs::copy(hosts_path, snapshot_path).map(|_| ())
-}
-
-fn write_staged_hosts_file(hosts_path: &Path, staged_path: &Path, content: &str) -> io::Result<()> {
-    maybe_fail_hosts_write_step(HostsWriteFailStep::StageWrite)?;
-    fs::write(staged_path, content)?;
-    // Copy the original file's permissions onto the staged file so replacement
-    // does not silently change the hosts file mode bits.
-    if let Ok(meta) = fs::metadata(hosts_path) {
-        let _ = fs::set_permissions(staged_path, meta.permissions());
-    }
-    Ok(())
-}
-
-#[cfg(target_os = "windows")]
-fn replace_hosts_file(staged_path: &Path, hosts_path: &Path) -> Result<(), HostsReplaceError> {
-    match fs::rename(staged_path, hosts_path) {
-        Ok(()) => Ok(()),
-        Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-            fs::remove_file(hosts_path).map_err(HostsReplaceError::NoMutation)?;
-            fs::rename(staged_path, hosts_path).map_err(HostsReplaceError::NeedsRollback)
-        }
-        Err(error) => Err(HostsReplaceError::NoMutation(error)),
-    }
-}
-
-#[cfg(not(target_os = "windows"))]
-fn replace_hosts_file(staged_path: &Path, hosts_path: &Path) -> Result<(), HostsReplaceError> {
-    fs::rename(staged_path, hosts_path).map_err(HostsReplaceError::NoMutation)
-}
-
-fn restore_hosts_snapshot(hosts_path: &Path, snapshot_path: &Path) -> io::Result<()> {
-    maybe_fail_hosts_write_step(HostsWriteFailStep::RollbackRestore)?;
-    restore_snapshot_file(snapshot_path, hosts_path)
-}
-
-#[cfg(target_os = "windows")]
-fn restore_snapshot_file(snapshot_path: &Path, hosts_path: &Path) -> io::Result<()> {
-    match fs::rename(snapshot_path, hosts_path) {
-        Ok(()) => Ok(()),
-        Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-            fs::remove_file(hosts_path)?;
-            fs::rename(snapshot_path, hosts_path)
-        }
-        Err(error) => Err(error),
-    }
-}
-
-#[cfg(not(target_os = "windows"))]
-fn restore_snapshot_file(snapshot_path: &Path, hosts_path: &Path) -> io::Result<()> {
-    fs::rename(snapshot_path, hosts_path)
-}
-
-fn hosts_transaction_temp_path(hosts_path: &Path, marker: &str) -> PathBuf {
-    let dir = hosts_path.parent().unwrap_or(Path::new("."));
-    let target_name = hosts_path
-        .file_name()
-        .and_then(|name| name.to_str())
-        .unwrap_or("hosts");
-    let pid = std::process::id();
-    let nanos = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|duration| duration.as_nanos())
-        .unwrap_or(0);
-    dir.join(format!(
-        ".{target_name}.focustime.{pid}.{nanos}.{marker}.tmp"
-    ))
-}
-
-fn remove_file_if_exists(path: &Path) -> io::Result<()> {
-    match fs::remove_file(path) {
-        Ok(()) => Ok(()),
-        Err(error) if error.kind() == io::ErrorKind::NotFound => Ok(()),
-        Err(error) => Err(error),
-    }
-}
-
-fn cleanup_temp_file_best_effort(path: &Path) {
-    if let Err(error) = remove_file_if_exists(path) {
-        eprintln!(
-            "Warning: failed to remove hosts temp file `{}`: {error}",
-            path.display()
-        );
-    }
-}
-
-fn maybe_fail_hosts_write_step(step: HostsWriteFailStep) -> io::Result<()> {
-    #[cfg(test)]
-    {
-        let should_fail = TEST_HOSTS_WRITE_FAIL_STEPS.with(|slot| {
-            let mut configured = slot.borrow_mut();
-            if configured.first().copied() == Some(step) {
-                configured.remove(0);
-                true
-            } else {
-                false
-            }
-        });
-        if should_fail {
-            return Err(io::Error::other(format!(
-                "simulated hosts write failure at {step:?}"
-            )));
-        }
-    }
-    #[cfg(not(test))]
-    {
-        let _ = step;
-    }
-
-    Ok(())
-}
-
-/// Flush the OS DNS cache so /etc/hosts changes take effect immediately.
-/// Best-effort: failures are silently ignored.
-fn flush_dns_cache() {
-    #[cfg(target_os = "macos")]
-    {
-        // Flush mDNSResponder cache (macOS 10.10.4+)
-        let _ = Command::new("dscacheutil").arg("-flushcache").status();
-        let _ = Command::new("killall")
-            .args(["-HUP", "mDNSResponder"])
-            .status();
-    }
-    #[cfg(target_os = "linux")]
-    {
-        // systemd-resolved
-        let _ = Command::new("systemd-resolve")
-            .arg("--flush-caches")
-            .status();
-        // nscd (older systems)
-        let _ = Command::new("nscd").args(["-i", "hosts"]).status();
-    }
-    #[cfg(target_os = "windows")]
-    {
-        let _ = Command::new("ipconfig").arg("/flushdns").status();
-    }
-}
-
 #[cfg(test)]
-mod tests {
-    use crate::blocker::*;
-    use std::path::PathBuf;
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    fn temp_hosts_path(label: &str) -> PathBuf {
-        let unique = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("system clock should be after unix epoch")
-            .as_nanos();
-        std::env::temp_dir().join(format!("focustime-hosts-rollback-{label}-{unique}.tmp"))
-    }
-
-    #[test]
-    fn strip_unterminated_start_marker_leaves_content_unchanged() {
-        // A lone start marker without an end marker must not drop any content.
-        let input = "127.0.0.1 localhost\n# focustime-block-start\n127.0.0.1 example.com\n";
-        assert_eq!(SiteBlocker::strip_block_section(input), input);
-    }
-
-    #[test]
-    fn strip_empty_string() {
-        assert_eq!(SiteBlocker::strip_block_section(""), "");
-    }
-
-    #[test]
-    fn strip_no_block_section_is_unchanged() {
-        let input = "127.0.0.1 localhost\n::1 localhost\n";
-        assert_eq!(SiteBlocker::strip_block_section(input), input);
-    }
-
-    #[test]
-    fn strip_removes_block_section() {
-        let input = "127.0.0.1 localhost\n# focustime-block-start\n127.0.0.1 example.com\n# focustime-block-end\n::1 localhost\n";
-        let expected = "127.0.0.1 localhost\n::1 localhost\n";
-        assert_eq!(SiteBlocker::strip_block_section(input), expected);
-    }
-
-    #[test]
-    fn strip_removes_block_section_at_end_of_file() {
-        let input = "127.0.0.1 localhost\n# focustime-block-start\n127.0.0.1 example.com\n# focustime-block-end\n";
-        let expected = "127.0.0.1 localhost\n";
-        assert_eq!(SiteBlocker::strip_block_section(input), expected);
-    }
-
-    #[test]
-    fn strip_multiple_sites_in_section() {
-        let input = "before\n# focustime-block-start\n127.0.0.1 a.com\n127.0.0.1 b.com\n# focustime-block-end\nafter\n";
-        let expected = "before\nafter\n";
-        assert_eq!(SiteBlocker::strip_block_section(input), expected);
-    }
-
-    #[test]
-    fn add_site_normalizes_and_deduplicates() {
-        let mut b = SiteBlocker::new();
-        b.add_site("  Example.COM  ".to_string());
-        b.add_site("example.com".to_string());
-        assert_eq!(b.sites, vec!["example.com"]);
-    }
-
-    #[test]
-    fn add_site_ignores_empty() {
-        let mut b = SiteBlocker::new();
-        b.add_site("   ".to_string());
-        assert!(b.sites.is_empty());
-    }
-
-    #[test]
-    fn add_site_strips_scheme_and_path() {
-        let mut b = SiteBlocker::new();
-        b.add_site("https://example.com/some/path?q=1".to_string());
-        assert_eq!(b.sites, vec!["example.com"]);
-    }
-
-    #[test]
-    fn add_site_strips_numeric_port() {
-        let mut b = SiteBlocker::new();
-        b.add_site("https://example.com:443/some/path".to_string());
-        assert_eq!(b.sites, vec!["example.com"]);
-    }
-
-    #[test]
-    fn add_site_rejects_multiple_hostnames() {
-        let mut b = SiteBlocker::new();
-        b.add_site("example.com other.com".to_string());
-        assert!(b.sites.is_empty());
-    }
-
-    #[test]
-    fn add_site_rejects_invalid_characters() {
-        let mut b = SiteBlocker::new();
-        b.add_site("exam_ple.com".to_string());
-        assert!(b.sites.is_empty());
-    }
-
-    #[test]
-    fn add_site_accepts_wildcard_rules() {
-        let mut b = SiteBlocker::new();
-        b.add_site("*.Docs.Example.com".to_string());
-        assert_eq!(b.sites, vec!["*.docs.example.com"]);
-    }
-
-    #[test]
-    fn add_site_rejects_mid_label_wildcard() {
-        let mut b = SiteBlocker::new();
-        b.add_site("foo*bar.example.com".to_string());
-        assert!(b.sites.is_empty());
-    }
-
-    #[test]
-    fn bulk_add_accepts_comma_and_newline_separators() {
-        let mut b = SiteBlocker::new();
-        let result = b.add_sites_from_input("example.com, github.com\nhttps://rust-lang.org/docs");
-        assert_eq!(
-            result.added,
-            vec!["example.com", "github.com", "rust-lang.org"]
-        );
-        assert!(result.duplicates.is_empty());
-        assert!(result.invalid.is_empty());
-        assert_eq!(b.sites, vec!["example.com", "github.com", "rust-lang.org"]);
-    }
-
-    #[test]
-    fn bulk_add_reports_duplicates_and_invalid_entries() {
-        let mut b = SiteBlocker::new();
-        let result = b.add_sites_from_input("github.com, bad host, exam_ple.com, github.com");
-        assert_eq!(result.added, vec!["github.com"]);
-        assert_eq!(result.duplicates, vec!["github.com"]);
-        assert_eq!(
-            result.invalid,
-            vec![
-                InvalidSiteInput {
-                    input: "bad host".to_string(),
-                    reason: SiteValidationError::ContainsWhitespace,
-                },
-                InvalidSiteInput {
-                    input: "exam_ple.com".to_string(),
-                    reason: SiteValidationError::InvalidCharacter,
-                }
-            ]
-        );
-    }
-
-    #[test]
-    fn edit_site_updates_selected_entry() {
-        let mut b = SiteBlocker::new();
-        b.add_site("a.com".to_string());
-        let result = b.edit_site_from_input(0, "https://news.ycombinator.com:443/newest");
-        assert_eq!(
-            result,
-            EditSiteResult::Updated {
-                old: "a.com".to_string(),
-                new: "news.ycombinator.com".to_string()
-            }
-        );
-        assert_eq!(b.sites, vec!["news.ycombinator.com"]);
-    }
-
-    #[test]
-    fn edit_site_rejects_duplicate_hostname() {
-        let mut b = SiteBlocker::new();
-        b.add_site("a.com".to_string());
-        b.add_site("b.com".to_string());
-        let result = b.edit_site_from_input(0, "b.com");
-        assert_eq!(
-            result,
-            EditSiteResult::Duplicate {
-                hostname: "b.com".to_string()
-            }
-        );
-        assert_eq!(b.sites, vec!["a.com", "b.com"]);
-    }
-
-    #[test]
-    fn edit_site_rejects_multiple_hostnames() {
-        let mut b = SiteBlocker::new();
-        b.add_site("a.com".to_string());
-        let result = b.edit_site_from_input(0, "a.com, b.com");
-        assert_eq!(
-            result,
-            EditSiteResult::Invalid(InvalidSiteInput {
-                input: "a.com, b.com".to_string(),
-                reason: SiteValidationError::MultipleHostnames,
-            })
-        );
-    }
-
-    #[test]
-    fn wildcard_rule_matches_subdomains_only() {
-        assert!(domain_rule_matches_host("*.example.com", "www.example.com"));
-        assert!(domain_rule_matches_host("*.example.com", "a.b.example.com"));
-        assert!(!domain_rule_matches_host("*.example.com", "example.com"));
-    }
-
-    #[test]
-    fn wildcard_rule_respects_label_boundaries() {
-        assert!(!domain_rule_matches_host("*.example.com", "badexample.com"));
-        assert!(!domain_rule_matches_host(
-            "*.example.com",
-            "example.com.bad"
-        ));
-    }
-
-    #[test]
-    fn normalize_domain_rule_supports_wildcards_after_url_prefix_stripping() {
-        assert_eq!(
-            normalize_domain_rule("https://*.example.com/path").unwrap(),
-            "*.example.com"
-        );
-    }
-
-    #[test]
-    fn normalize_domain_rule_canonicalizes_dotted_forms() {
-        assert_eq!(
-            normalize_domain_rule(".Example.com").unwrap(),
-            "*.example.com"
-        );
-        assert_eq!(
-            normalize_domain_rule("example.com.").unwrap(),
-            "example.com"
-        );
-        assert_eq!(
-            normalize_domain_rule("https://.Example.com./path").unwrap(),
-            "*.example.com"
-        );
-        assert_eq!(
-            normalize_domain_rule(".com").unwrap_err(),
-            SiteValidationError::InvalidLabel
-        );
-    }
-
-    #[test]
-    fn normalize_domain_host_canonicalizes_root_dot_forms() {
-        assert_eq!(
-            normalize_domain_host(".WWW.Example.com.").unwrap(),
-            "www.example.com"
-        );
-    }
-
-    #[test]
-    fn wildcard_matching_handles_punycode_and_dotted_forms() {
-        assert!(domain_rule_matches_host(
-            ".xn--bcher-kva.example",
-            "shop.xn--bcher-kva.example."
-        ));
-        assert!(domain_rule_matches_host(
-            "*.example.com.",
-            "service.example.com"
-        ));
-        assert!(!domain_rule_matches_host("*.example.com.", "example.com."));
-    }
-
-    #[test]
-    fn exact_rule_matches_only_exact_hostname() {
-        assert!(domain_rule_matches_host("example.com", "example.com"));
-        assert!(!domain_rule_matches_host("example.com", "www.example.com"));
-    }
-
-    #[test]
-    fn strip_out_of_order_markers_leaves_content_unchanged() {
-        // End marker before start marker: treat as corrupt, return unchanged.
-        let input = "127.0.0.1 localhost\n# focustime-block-end\n# focustime-block-start\nafter\n";
-        assert_eq!(SiteBlocker::strip_block_section(input), input);
-    }
-
-    #[test]
-    fn strip_marker_with_trailing_content_leaves_unchanged() {
-        // Markers that appear as substrings of longer lines must not be treated
-        // as valid markers; the whole file should be returned untouched.
-        let input = "127.0.0.1 localhost\n# focustime-block-start\n127.0.0.1 example.com\n# focustime-block-end extra\n::1 localhost\n";
-        assert_eq!(SiteBlocker::strip_block_section(input), input);
-    }
-
-    #[test]
-    fn strip_preserves_crlf_line_endings() {
-        let input = "127.0.0.1 localhost\r\n# focustime-block-start\r\n127.0.0.1 example.com\r\n# focustime-block-end\r\n::1 localhost\r\n";
-        let expected = "127.0.0.1 localhost\r\n::1 localhost\r\n";
-        assert_eq!(SiteBlocker::strip_block_section(input), expected);
-    }
-
-    #[test]
-    fn remove_site_by_index() {
-        let mut b = SiteBlocker::new();
-        b.add_site("a.com".to_string());
-        b.add_site("b.com".to_string());
-        let removed = b.remove_site(0);
-        assert_eq!(removed.as_deref(), Some("a.com"));
-        assert_eq!(b.sites, vec!["b.com"]);
-    }
-
-    #[test]
-    fn remove_site_out_of_bounds_is_safe() {
-        let mut b = SiteBlocker::new();
-        b.add_site("a.com".to_string());
-        assert!(b.remove_site(5).is_none()); // should not panic
-        assert_eq!(b.sites.len(), 1);
-    }
-
-    #[test]
-    fn hosts_file_diagnostics_reports_read_and_write_success_for_temp_file() {
-        let unique = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("system clock should be after unix epoch")
-            .as_nanos();
-        let path = std::env::temp_dir().join(format!("focustime-hosts-diagnostics-{unique}.tmp"));
-        fs::write(&path, "127.0.0.1 localhost\n").expect("temp hosts file should be writable");
-
-        let diagnostics = hosts_file_diagnostics_for(&path);
-
-        assert!(diagnostics.can_read());
-        assert!(diagnostics.can_write());
-        assert!(diagnostics.read_error.is_none());
-        assert!(diagnostics.write_error.is_none());
-        let _ = fs::remove_file(path);
-    }
-
-    #[test]
-    fn hosts_file_diagnostics_reports_missing_file_errors() {
-        let unique = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("system clock should be after unix epoch")
-            .as_nanos();
-        let path =
-            std::env::temp_dir().join(format!("focustime-hosts-diagnostics-missing-{unique}.tmp"));
-        let _ = fs::remove_file(&path);
-
-        let diagnostics = hosts_file_diagnostics_for(&path);
-
-        assert!(!diagnostics.can_read());
-        assert!(diagnostics.read_error.is_some());
-        #[cfg(target_os = "windows")]
-        {
-            assert!(!diagnostics.can_write());
-            assert!(diagnostics.write_error.is_some());
-        }
-        #[cfg(not(target_os = "windows"))]
-        {
-            assert!(diagnostics.can_write());
-            assert!(diagnostics.write_error.is_none());
-        }
-    }
-
-    #[test]
-    fn preview_block_reports_next_section_and_change() {
-        let mut blocker = SiteBlocker::new();
-        blocker.add_site("example.com".to_string());
-        let original = "127.0.0.1 localhost\n";
-
-        let preview = blocker.preview_from_hosts_content("hosts", original, BlockingIntent::Block);
-
-        assert_eq!(preview.action, BlockingPreviewAction::Block);
-        assert!(preview.would_change);
-        assert_eq!(preview.current_section, None);
-        assert_eq!(preview.effective_blocked_sites, vec!["example.com"]);
-        let section = preview
-            .next_section
-            .as_deref()
-            .expect("block preview should include next section");
-        assert!(section.contains("# focustime-block-start"));
-        assert!(section.contains("127.0.0.1 example.com"));
-        assert!(section.contains("::1 www.example.com"));
-        assert_eq!(preview.section_for_display(), Some(section));
-    }
-
-    #[test]
-    fn preview_block_skips_wildcard_entries_for_hosts_backend() {
-        let mut blocker = SiteBlocker::new();
-        blocker.add_site("*.example.com".to_string());
-        blocker.add_site("api.example.com".to_string());
-        let original = "127.0.0.1 localhost\n";
-
-        let preview = blocker.preview_from_hosts_content("hosts", original, BlockingIntent::Block);
-        let section = preview
-            .next_section
-            .as_deref()
-            .expect("block preview should include next section");
-
-        assert_eq!(preview.effective_blocked_sites, vec!["api.example.com"]);
-        assert!(!section.contains("*.example.com"));
-        assert!(section.contains("127.0.0.1 api.example.com"));
-    }
-
-    #[test]
-    fn preview_unblock_reports_current_section_and_change() {
-        let blocker = SiteBlocker::new();
-        let original = "127.0.0.1 localhost\n# focustime-block-start\n127.0.0.1 example.com\n# focustime-block-end\n";
-
-        let preview =
-            blocker.preview_from_hosts_content("hosts", original, BlockingIntent::Unblock);
-
-        assert_eq!(preview.action, BlockingPreviewAction::Unblock);
-        assert!(preview.would_change);
-        assert!(preview.next_section.is_none());
-        let section = preview
-            .current_section
-            .as_deref()
-            .expect("unblock preview should include current section");
-        assert!(section.contains("# focustime-block-start"));
-        assert!(section.contains("# focustime-block-end"));
-        assert_eq!(preview.section_for_display(), Some(section));
-    }
-
-    #[test]
-    fn preview_unblock_reports_all_existing_sections() {
-        let blocker = SiteBlocker::new();
-        let original = concat!(
-            "127.0.0.1 localhost\n",
-            "# focustime-block-start\n",
-            "127.0.0.1 example.com\n",
-            "# focustime-block-end\n",
-            "# focustime-block-start\n",
-            "127.0.0.1 github.com\n",
-            "# focustime-block-end\n",
-        );
-
-        let preview =
-            blocker.preview_from_hosts_content("hosts", original, BlockingIntent::Unblock);
-        let section = preview
-            .current_section
-            .as_deref()
-            .expect("unblock preview should include all current sections");
-
-        assert_eq!(section.matches("# focustime-block-start").count(), 2);
-        assert!(section.contains("127.0.0.1 example.com"));
-        assert!(section.contains("127.0.0.1 github.com"));
-    }
-
-    #[test]
-    fn preview_block_no_change_when_hosts_already_match() {
-        let mut blocker = SiteBlocker::new();
-        blocker.add_site("example.com".to_string());
-        let original = blocker
-            .build_blocked_hosts_content("127.0.0.1 localhost\n", &["example.com".to_string()]);
-
-        let preview = blocker.preview_from_hosts_content("hosts", &original, BlockingIntent::Block);
-
-        assert_eq!(preview.action, BlockingPreviewAction::NoChange);
-        assert!(!preview.would_change);
-        assert!(preview.next_section.is_some());
-        assert!(preview.current_section.is_some());
-    }
-
-    #[test]
-    fn block_rolls_back_to_original_after_post_replace_failure() {
-        let path = temp_hosts_path("block-post-replace");
-        let original = "127.0.0.1 localhost\n::1 localhost\n";
-        fs::write(&path, original).expect("temp hosts file should be writable");
-
-        let mut blocker = SiteBlocker::new();
-        blocker.add_site("example.com".to_string());
-        set_test_hosts_write_fail_steps(&[HostsWriteFailStep::AfterReplace]);
-
-        let result = blocker.apply_hosts_block_to_path(&path);
-        set_test_hosts_write_fail_steps(&[]);
-
-        assert!(result.is_err());
-        let restored = fs::read_to_string(&path).expect("hosts file should remain readable");
-        assert_eq!(restored, original);
-        assert!(!restored.contains(BLOCK_MARKER_START));
-        assert!(!restored.contains(BLOCK_MARKER_END));
-
-        let _ = fs::remove_file(path);
-    }
-
-    #[test]
-    fn block_failure_before_replace_keeps_original_content() {
-        let path = temp_hosts_path("block-stage-write");
-        let original = "127.0.0.1 localhost\n::1 localhost\n";
-        fs::write(&path, original).expect("temp hosts file should be writable");
-
-        let mut blocker = SiteBlocker::new();
-        blocker.add_site("example.com".to_string());
-        set_test_hosts_write_fail_steps(&[HostsWriteFailStep::StageWrite]);
-
-        let result = blocker.apply_hosts_block_to_path(&path);
-        set_test_hosts_write_fail_steps(&[]);
-
-        assert!(result.is_err());
-        let content = fs::read_to_string(&path).expect("hosts file should remain readable");
-        assert_eq!(content, original);
-        assert!(!content.contains(BLOCK_MARKER_START));
-        assert!(!content.contains(BLOCK_MARKER_END));
-
-        let _ = fs::remove_file(path);
-    }
-
-    #[test]
-    fn unblock_rolls_back_to_complete_section_after_post_replace_failure() {
-        let path = temp_hosts_path("unblock-post-replace");
-        let original = concat!(
-            "127.0.0.1 localhost\n",
-            "# focustime-block-start\n",
-            "127.0.0.1 example.com\n",
-            "::1 example.com\n",
-            "# focustime-block-end\n",
-            "::1 localhost\n",
-        );
-        fs::write(&path, original).expect("temp hosts file should be writable");
-
-        let blocker = SiteBlocker::new();
-        set_test_hosts_write_fail_steps(&[HostsWriteFailStep::AfterReplace]);
-
-        let result = blocker.remove_hosts_block_from_path(&path);
-        set_test_hosts_write_fail_steps(&[]);
-
-        assert!(result.is_err());
-        let restored = fs::read_to_string(&path).expect("hosts file should remain readable");
-        assert_eq!(restored, original);
-        assert_eq!(restored.matches(BLOCK_MARKER_START).count(), 1);
-        assert_eq!(restored.matches(BLOCK_MARKER_END).count(), 1);
-
-        let _ = fs::remove_file(path);
-    }
-
-    #[test]
-    fn rollback_failure_is_reported_when_restore_step_fails() {
-        let path = temp_hosts_path("rollback-restore-error");
-        let original = "127.0.0.1 localhost\n::1 localhost\n";
-        fs::write(&path, original).expect("temp hosts file should be writable");
-
-        let mut blocker = SiteBlocker::new();
-        blocker.add_site("example.com".to_string());
-        set_test_hosts_write_fail_steps(&[
-            HostsWriteFailStep::AfterReplace,
-            HostsWriteFailStep::RollbackRestore,
-        ]);
-
-        let result = blocker.apply_hosts_block_to_path(&path);
-        set_test_hosts_write_fail_steps(&[]);
-
-        let error = result.expect_err("rollback should fail when restore step is injected to fail");
-        assert!(error.to_string().contains("rollback failed"));
-        let content = fs::read_to_string(&path).expect("hosts file should remain readable");
-        assert!(content.contains(BLOCK_MARKER_START));
-        assert!(content.contains(BLOCK_MARKER_END));
-
-        let _ = fs::remove_file(path);
-    }
-}
+mod tests;
